+++ Gruntfile.js
@@ -58,7 +58,7 @@ module.exports = function( grunt ) {
 
 				// Exclude specified modules if the module matching the key is removed
 				removeWith: {
-					ajax: [ "manipulation/_evalUrl", "event/ajax" ],
+					ajax: [ "manipulation/_evalUrl", "deprecated/ajax-event-alias" ],
 					callbacks: [ "deferred" ],
 					css: [ "effects", "dimensions", "offset" ],
 					"css/showHide": [ "effects" ],
@@ -66,6 +66,7 @@ module.exports = function( grunt ) {
 						remove: [ "ajax", "effects", "queue", "core/ready" ],
 						include: [ "core/ready-no-deferred" ]
 					},
+					event: [ "deprecated/ajax-event-alias", "deprecated/event" ],
 					sizzle: [ "css/hiddenVisibleSelectors", "effects/animatedSelector" ]
 				}
 			}
