+++ src/deprecated/ajax-event-alias.js
@@ -1,11 +1,11 @@
 define( [
 	"../core",
+	"../ajax",
 	"../event"
 ], function( jQuery ) {
 
 "use strict";
 
-// Attach a bunch of functions for handling common AJAX events
 jQuery.each( [
 	"ajaxStart",
 	"ajaxStop",
