+++ src/jquery.js
@@ -23,7 +23,6 @@ define( [
 	"./ajax/script",
 	"./ajax/jsonp",
 	"./ajax/load",
-	"./event/ajax",
 	"./effects",
 	"./effects/animatedSelector",
 	"./offset",
